+++ packages/ir-sdk/fern/apis/ir-types-latest/changelog/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [v54.0.1] - 2024-01-10
+
+- Feature: OAuthAccessTokenRequestProperties now pulls in additional request properties from the OAuth
+  getToken endpoint to support custom OAuth schemas.
+
 ## [v54.0.0] - 2024-12-11
 
 - Break: The HttpResponse type in the IR now supports bytes responses. This is useful for different languages -